Add "support" (I think) for thin provisioning pools.
Segments which are of type "thin-pool" are used to define thin provisioning
pools, which are composed of other ("normal") logical volumes. Segments of
type "thin" indicate they are stored in a thin pool. Neither type of segment
actually exists as a standalone physical segment that can/should be manipulated
by this utility.

My knowledge of PHP is limited to what I read in these scripts ;) but this
commit squelches output of thin entities by setting their segment count to 0
and appears to achieve the intended results.

Without this fix, dump.php would generate some entries for thin volumes and
rearrange.php would abort because no actual storage information appears for
them. After the fix, thin volumes do not appear in any output files and it
is possible to proceed with defragmentation.